feat 공통UI Input 추가
