+++ .github/ISSUE_TEMPLATE/file_change.md
@@ -86,3 +86,29 @@ assignees: ""
 ## 📝 변경 이유
 
 ### <!-- 파일이 수정된 이유가 있다면 작성 부탁 드립니다. 따로 없으면 작성 안해도 무방합니다. -->
+
+### 2025.10.17
+
+#### 🔹 공통 UI 컴포넌트
+
+- 공통UI(Input) 추가
+
+#### 🔹 정적 리소스 관련
+
+-
+
+### 위치
+
+- src/shared/lib/utils
+- src/shared/ui/button
+- src/shared/ui/input
+
+## 📌 파일 변경 내용 및 위치
+
+- 공통 UI에 사용될 Input 추가
+
+## 📝 변경 이유
+
+- 프로젝트 전반에 사용될 input 이기 때문에 추가
+
+### <!-- 파일이 수정된 이유가 있다면 작성 부탁 드립니다. 따로 없으면 작성 안해도 무방합니다. -->
