+++ src/assets/icons/button/index.ts
@@ -1,0 +1,2 @@
+export * from "./closeButton";
+export * from "./leftButton";
