+++ src/assets/icons/home/index.ts
@@ -1,0 +1,4 @@
+export * from "./arrowUp";
+export * from "./homeIcon";
+export * from "./person";
+export * from "./search";
