+++ src/assets/icons/logo/index.ts
@@ -1,0 +1,2 @@
+export * from "./logIcon";
+export * from "./pinHouseLogo";
