+++ src/assets/icons/onboarding/index.ts
@@ -1,0 +1,4 @@
+export * from "./TaskList";
+export * from "./house";
+export * from "./mapPin";
+export * from "./propertyBuilding";
