--- src/shared/ui/button/Button.tsx
@@ -1,35 +1,0 @@
-"use client";
-import { cva, type VariantProps } from "class-variance-authority";
-import { ButtonHTMLAttributes } from "react";
-import { cn } from "../../lib/utils";
-
-const buttonVariants = cva(
-  "inline-flex items-center justify-center rounded-2xl font-medium transition-all focus:outline-none disabled:opacity-50 disabled:pointer-events-none",
-  {
-    variants: {
-      variant: {
-        solid:
-          "g-primary text-white hover:bg-primary/90 active:scale-[0.98] text-white bg-button-light rounded-3xl w-full",
-        outline: "border border-gray-300 text-gray-800 hover:bg-gray-50 active:scale-[0.98]",
-        ghost: "bg-transparent text-gray-700 hover:bg-gray-100 active:scale-[0.98]",
-      },
-      size: {
-        sm: "h-8 px-3 text-sm",
-        md: "h-10 px-4 text-base",
-        lg: "h-12 px-6 text-lg",
-      },
-    },
-    defaultVariants: {
-      variant: "solid",
-      size: "md",
-    },
-  }
-);
-
-interface ButtonProps
-  extends ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {}
-
-export function Button({ className, variant, size, ...props }: ButtonProps) {
-  return <button className={cn(buttonVariants({ variant, size }), className)} {...props} />;
-}
