+++ src/shared/ui/button/deafult/button.variants.ts
@@ -1,19 +1,19 @@
-import { cva, VariantProps } from "class-variance-authority";
-import { ButtonHTMLAttributes } from "react";
+import { cva } from "class-variance-authority";
 
-const buttonVariants = cva(
+export const buttonVariants = cva(
   "inline-flex items-center justify-center rounded-2xl font-medium transition-all focus:outline-none disabled:opacity-50 disabled:pointer-events-none",
   {
     variants: {
       variant: {
-        solid: "bg-primary text-white hover:bg-primary/90 active:scale-[0.98]",
+        solid:
+          "g-primary text-white hover:bg-primary/90 active:scale-[0.98] text-white bg-button-light rounded-3xl",
         outline: "border border-gray-300 text-gray-800 hover:bg-gray-50 active:scale-[0.98]",
         ghost: "bg-transparent text-gray-700 hover:bg-gray-100 active:scale-[0.98]",
       },
       size: {
         sm: "h-8 px-3 text-sm",
         md: "h-10 px-4 text-base",
-        lg: "h-12 px-6 text-lg",
+        lg: "h-12 w-full px-5 text-lg",
       },
     },
     defaultVariants: {
@@ -22,7 +22,3 @@ const buttonVariants = cva(
     },
   }
 );
-
-export interface ButtonProps
-  extends ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {}
