--- src/shared/ui/button/index.ts
@@ -1,3 +1,0 @@
-export { Button } from "./Button";
-export type { ButtonProps, ButtonVariant, ButtonSize } from "./types";
-
