+++ src/shared/ui/input/deafult/index.ts
@@ -1,0 +1,2 @@
+export * from "./input";
+export type { InputProps } from "./type";
